Add cost computation from logits
The log required for the cost computation is computed using the logits and not softmax output of the last layer. This is done for escaping the numerical stability issues caused due to softmax (often seen as NaN) during backpropagation. No division action is performed during this operation. And the logit is normalized(shifted) inorder to avoid numerical overflow.